+++ package.json
@@ -21,7 +21,6 @@
     "vue-axios": "^2.1.4",
     "vuex": "^3.1.1",
     "vuex-persistedstate": "^2.5.4"
-
   },
   "devDependencies": {
     "@babel/core": "~7.2",
