+++ src/backend/jsonCocktail.json
@@ -1,8 +1,8 @@
-﻿[
-  {
+﻿{
+  "data": [{
     "id": 0,
     "name": "Цвета осени",
-    "src": "src/store/static/assets/image/icon_1537758161-Black-Cherry__highres.jpg",
+    "src": "/assets/image/icon_1537758161-Black-Cherry__highres.jpg",
     "alco": 40,
     "ingredient": "фрукты, овощи",
     "price": "450",
@@ -10,11 +10,10 @@
     "taste": "Фруктовые",
     "color": "Желтые",
     "alcoStr": "Крепкие"
-  },
-  {
+  }, {
     "id": 1,
     "name": "Розовый сад",
-    "src": "src/store/static/assets/image/Rose_garden.jpg",
+    "src": "/assets/image/Rose_garden.jpg",
     "alco": "14",
     "ingredient": "фрукты, овощи",
     "price": "150",
@@ -22,11 +21,11 @@
     "taste": "Фруктовые",
     "color": "Красные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 2,
     "name": "Королевский лимонад",
-    "src": "src/store/static/assets/image/icon_1537766175-Lemonade-Royale__highres.jpg",
+    "src": "/assets/image/icon_1537766175-Lemonade-Royale__highres.jpg",
     "alco": "18",
     "ingredient": "фрукты, овощи",
     "price": "550",
@@ -34,11 +33,11 @@
     "taste": "Фруктовые",
     "color": "Красные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 3,
     "name": "Пойдем домой",
-    "src": "src/store/static/assets/image/icon_1538166598-Take_me_home-HiRes.jpg",
+    "src": "/assets/image/icon_1538166598-Take_me_home-HiRes.jpg",
     "alco": "18",
     "ingredient": "фрукты, овощи фруктыфруктыфруктыфрукты",
     "price": "450",
@@ -46,11 +45,10 @@
     "taste": "Фруктовые",
     "color": "Красные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+  }, {
     "id": 4,
     "name": "Бока-чика",
-    "src": "src/store/static/assets/image/icon_1540546526-Boca-Chica__highres.jpg",
+    "src": "/assets/image/icon_1540546526-Boca-Chica__highres.jpg",
     "alco": "12",
     "ingredient": "фрукты, овощи",
     "price": "150",
@@ -58,11 +56,11 @@
     "taste": "Фруктовые",
     "color": "Зеленые",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 5,
     "name": "Альма",
-    "src": "src/store/static/assets/image/icon_1551358867-Alma_800.jpg",
+    "src": "/assets/image/icon_1551358867-Alma_800.jpg",
     "alco": "17",
     "ingredient": "фрукты, овощи",
     "price": "350",
@@ -70,11 +68,11 @@
     "taste": "Фруктовые",
     "color": "Прозрачные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 6,
     "name": "Тай",
-    "src": "src/store/static/assets/image/icon_1556010828-Thai-Sling__highres.jpg",
+    "src": "/assets/image/icon_1556010828-Thai-Sling__highres.jpg",
     "alco": "18",
     "ingredient": "фрукты, овощи",
     "price": "350",
@@ -82,11 +80,11 @@
     "taste": "Фруктовые",
     "color": "Красные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 7,
     "name": "Олдфэшн",
-    "src": "src/store/static/assets/image/icon_1556121518-Old-fashioned-w800px.jpg",
+    "src": "/assets/image/icon_1556121518-Old-fashioned-w800px.jpg",
     "alco": "40",
     "ingredient": "фрукты, овощи",
     "price": "350",
@@ -94,11 +92,11 @@
     "taste": "Горькие",
     "color": "Желтые",
     "alcoStr": "Крепкие"
-  },
-  {
+
+  }, {
     "id": 8,
     "name": "Пармигиано",
-    "src": "src/store/static/assets/image/icon_1556461694-Parmigiano_sour_800.jpg",
+    "src": "/assets/image/icon_1556461694-Parmigiano_sour_800.jpg",
     "alco": "14",
     "ingredient": "фрукты, овощи",
     "price": "450",
@@ -106,11 +104,11 @@
     "taste": "Горькие",
     "color": "Желтые",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 9,
     "name": "Скользкий",
-    "src": "src/store/static/assets/image/icon_1556798272-Slippery-Nipple__highres.jpg",
+    "src": "/assets/image/icon_1556798272-Slippery-Nipple__highres.jpg",
     "alco": "15",
     "ingredient": "фрукты, овощи",
     "price": "450",
@@ -118,11 +116,11 @@
     "taste": "Горькие",
     "color": "Красные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 10,
     "name": "Северный",
-    "src": "src/store/static/assets/image/icon_1557228199-North-Sling__highres.jpg",
+    "src": "/assets/image/icon_1557228199-North-Sling__highres.jpg",
     "alco": "18",
     "ingredient": "фрукты, овощи",
     "price": "350",
@@ -130,11 +128,11 @@
     "taste": "Горькие",
     "color": "Красные",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+
+  }, {
     "id": 11,
     "name": "Летний пунш",
-    "src": "src/store/static/assets/image/icon_1557229223-Spring-Punch__highres.jpg",
+    "src": "/assets/image/icon_1557229223-Spring-Punch__highres.jpg",
     "alco": "0",
     "ingredient": "фрукты, овощи",
     "price": "250",
@@ -142,11 +140,11 @@
     "taste": "Фруктовые",
     "color": "Красные",
     "alcoStr": "Безалкогольные"
-  },
-  {
+
+  }, {
     "id": 12,
     "name": "Яблочный лимонад",
-    "src": "src/store/static/assets/image/icon_Apple-Lemonade__highres.jpg",
+    "src": "/assets/image/icon_Apple-Lemonade__highres.jpg",
     "alco": "0",
     "ingredient": "фрукты, овощи",
     "price": "250",
@@ -154,11 +152,11 @@
     "taste": "Фруктовые",
     "color": "Желтые",
     "alcoStr": "Безалкогольные"
-  },
-  {
+
+  }, {
     "id": 13,
     "name": "Мохито",
-    "src": "src/store/static/assets/image/icon_Bacardi_shooting_5_800.jpg",
+    "src": "/assets/image/icon_Bacardi_shooting_5_800.jpg",
     "alco": "39",
     "ingredient": "\u0411\u0435\u043B\u044B\u0439 \u0440\u043E\u043C Bacardi 50 \u043C\u043B, \u0421\u0430\u0445\u0430\u0440\u043D\u044B\u0439 \u0441\u0438\u0440\u043E\u043F 15\t\u043C\u043B, \u0421\u043E\u0434\u043E\u0432\u0430\u044F 100 \u043C\u043B, \u041B\u0430\u0439\u043C 80\u0433, \u041C\u044F\u0442\u0430 3\u0433, \u0414\u0440\u043E\u0431\u043B\u0435\u043D\u044B\u0439 \u043B\u0435\u0434 200\u0433",
     "price": "350",
@@ -166,11 +164,11 @@
     "taste": "Кислые",
     "color": "Прозрачные",
     "alcoStr": "Крепкие"
-  },
-  {
+
+  }, {
     "id": 14,
     "name": "Апероль Шприц",
-    "src": "src/store/static/assets/image/icon_1537863520-Aperol_spritz-HiRes.jpg",
+    "src": "/assets/image/icon_1537863520-Aperol_spritz-HiRes.jpg",
     "alco": "25",
     "ingredient": "фрукты, овощи",
     "price": "550",
@@ -178,11 +176,10 @@
     "taste": "Кислые",
     "color": "Желтые",
     "alcoStr": "Слабоалкогольные"
-  },
-  {
+  }, {
     "id": 15,
     "name": "???",
-    "src": "src/store/static/assets/image/icon_com_highres2.jpg",
+    "src": "/assets/image/icon_com_highres2.jpg",
     "alco": "78",
     "ingredient": "фрукты, овощи",
     "price": "500",
@@ -190,11 +187,11 @@
     "taste": "Фруктовые",
     "color": "Желтые",
     "alcoStr": "Крепкие"
-  },
-  {
+
+  }, {
     "id": 16,
     "name": "Клюквенный бурбон",
-    "src": "src/store/static/assets/image/icon_Cranberry-infused_bourbon-highres.jpg",
+    "src": "/assets/image/icon_Cranberry-infused_bourbon-highres.jpg",
     "alco": "38",
     "ingredient": "фрукты, овощи",
     "price": "400",
@@ -202,5 +199,5 @@
     "taste": "Фруктовые",
     "color": "Красные",
     "alcoStr": "Крепкие"
-  }
-]
+  }]
+}
