--- src/backend/static/assets/json/jsonCocktail.json
@@ -1,203 +1,0 @@
-﻿{
-  "data": [{
-    "id": 0,
-    "name": "Цвета осени",
-    "src": "src/backend/static/assets/image/icon_1537758161-Black-Cherry__highres.jpg",
-    "alco": 40,
-    "ingredient": "фрукты, овощи",
-    "price": "450",
-    "vol": "150",
-    "taste": "Фруктовые",
-    "color": "Желтые",
-    "alcoStr": "Крепкие"
-  }, {
-    "id": 1,
-    "name": "Розовый сад",
-    "src": "src/backend/static/assets/image/Rose_garden.jpg",
-    "alco": "14",
-    "ingredient": "фрукты, овощи",
-    "price": "150",
-    "vol": "50",
-    "taste": "Фруктовые",
-    "color": "Красные",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 2,
-    "name": "Королевский лимонад",
-    "src": "src/backend/static/assets/image/icon_1537766175-Lemonade-Royale__highres.jpg",
-    "alco": "18",
-    "ingredient": "фрукты, овощи",
-    "price": "550",
-    "vol": "150",
-    "taste": "Фруктовые",
-    "color": "Красные",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 3,
-    "name": "Пойдем домой",
-    "src": "src/backend/static/assets/image/icon_1538166598-Take_me_home-HiRes.jpg",
-    "alco": "18",
-    "ingredient": "фрукты, овощи фруктыфруктыфруктыфрукты",
-    "price": "450",
-    "vol": "150",
-    "taste": "Фруктовые",
-    "color": "Красные",
-    "alcoStr": "Слабоалкогольные"
-  }, {
-    "id": 4,
-    "name": "Бока-чика",
-    "src": "src/backend/static/assets/image/icon_1540546526-Boca-Chica__highres.jpg",
-    "alco": "12",
-    "ingredient": "фрукты, овощи",
-    "price": "150",
-    "vol": "50",
-    "taste": "Фруктовые",
-    "color": "Зеленые",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 5,
-    "name": "Альма",
-    "src": "src/backend/static/assets/image/icon_1551358867-Alma_800.jpg",
-    "alco": "17",
-    "ingredient": "фрукты, овощи",
-    "price": "350",
-    "vol": "250",
-    "taste": "Фруктовые",
-    "color": "Прозрачные",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 6,
-    "name": "Тай",
-    "src": "src/backend/static/assets/image/icon_1556010828-Thai-Sling__highres.jpg",
-    "alco": "18",
-    "ingredient": "фрукты, овощи",
-    "price": "350",
-    "vol": "250",
-    "taste": "Фруктовые",
-    "color": "Красные",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 7,
-    "name": "Олдфэшн",
-    "src": "src/backend/static/assets/image/icon_1556121518-Old-fashioned-w800px.jpg",
-    "alco": "40",
-    "ingredient": "фрукты, овощи",
-    "price": "350",
-    "vol": "100",
-    "taste": "Горькие",
-    "color": "Желтые",
-    "alcoStr": "Крепкие"
-
-  }, {
-    "id": 8,
-    "name": "Пармигиано",
-    "src": "src/backend/static/assets/image/icon_1556461694-Parmigiano_sour_800.jpg",
-    "alco": "14",
-    "ingredient": "фрукты, овощи",
-    "price": "450",
-    "vol": "200",
-    "taste": "Горькие",
-    "color": "Желтые",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 9,
-    "name": "Скользкий",
-    "src": "src/backend/static/assets/image/icon_1556798272-Slippery-Nipple__highres.jpg",
-    "alco": "15",
-    "ingredient": "фрукты, овощи",
-    "price": "450",
-    "vol": "50",
-    "taste": "Горькие",
-    "color": "Красные",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 10,
-    "name": "Северный",
-    "src": "src/backend/static/assets/image/icon_1557228199-North-Sling__highres.jpg",
-    "alco": "18",
-    "ingredient": "фрукты, овощи",
-    "price": "350",
-    "vol": "200",
-    "taste": "Горькие",
-    "color": "Красные",
-    "alcoStr": "Слабоалкогольные"
-
-  }, {
-    "id": 11,
-    "name": "Летний пунш",
-    "src": "src/backend/static/assets/image/icon_1557229223-Spring-Punch__highres.jpg",
-    "alco": "0",
-    "ingredient": "фрукты, овощи",
-    "price": "250",
-    "vol": "450",
-    "taste": "Фруктовые",
-    "color": "Красные",
-    "alcoStr": "Безалкогольные"
-
-  }, {
-    "id": 12,
-    "name": "Яблочный лимонад",
-    "src": "src/backend/static/assets/image/icon_Apple-Lemonade__highres.jpg",
-    "alco": "0",
-    "ingredient": "фрукты, овощи",
-    "price": "250",
-    "vol": "450",
-    "taste": "Фруктовые",
-    "color": "Желтые",
-    "alcoStr": "Безалкогольные"
-
-  }, {
-    "id": 13,
-    "name": "Мохито",
-    "src": "src/backend/static/assets/image/icon_Bacardi_shooting_5_800.jpg",
-    "alco": "39",
-    "ingredient": "\u0411\u0435\u043B\u044B\u0439 \u0440\u043E\u043C Bacardi 50 \u043C\u043B, \u0421\u0430\u0445\u0430\u0440\u043D\u044B\u0439 \u0441\u0438\u0440\u043E\u043F 15\t\u043C\u043B, \u0421\u043E\u0434\u043E\u0432\u0430\u044F 100 \u043C\u043B, \u041B\u0430\u0439\u043C 80\u0433, \u041C\u044F\u0442\u0430 3\u0433, \u0414\u0440\u043E\u0431\u043B\u0435\u043D\u044B\u0439 \u043B\u0435\u0434 200\u0433",
-    "price": "350",
-    "vol": "100",
-    "taste": "Кислые",
-    "color": "Прозрачные",
-    "alcoStr": "Крепкие"
-
-  }, {
-    "id": 14,
-    "name": "Апероль Шприц",
-    "src": "src/backend/static/assets/image/icon_1537863520-Aperol_spritz-HiRes.jpg",
-    "alco": "25",
-    "ingredient": "фрукты, овощи",
-    "price": "550",
-    "vol": "300",
-    "taste": "Кислые",
-    "color": "Желтые",
-    "alcoStr": "Слабоалкогольные"
-  }, {
-    "id": 15,
-    "name": "???",
-    "src": "src/backend/static/assets/image/icon_com_highres2.jpg",
-    "alco": "78",
-    "ingredient": "фрукты, овощи",
-    "price": "500",
-    "vol": "300",
-    "taste": "Фруктовые",
-    "color": "Желтые",
-    "alcoStr": "Крепкие"
-
-  }, {
-    "id": 16,
-    "name": "Клюквенный бурбон",
-    "src": "src/backend/static/assets/image/icon_Cranberry-infused_bourbon-highres.jpg",
-    "alco": "38",
-    "ingredient": "фрукты, овощи",
-    "price": "400",
-    "vol": "200",
-    "taste": "Фруктовые",
-    "color": "Красные",
-    "alcoStr": "Крепкие"
-  }]
-}
