+++ src/static/assets/json/dropdownItem.json
@@ -1,5 +1,6 @@
 {
-  "alco": [{
+  "alco": [
+    {
       "id": 1,
       "name": "Любые",
       "value": null
@@ -8,17 +9,20 @@
       "id": 2,
       "name": "Безалкогольные",
       "value": "Безалкогольные"
-    }, {
+    },
+    {
       "id": 3,
       "name": "Крепкие",
       "value": "Крепкие"
-    }, {
+    },
+    {
       "id": 4,
       "name": "Слабоалкогольные",
       "value": "Слабоалкогольные"
     }
   ],
-  "color": [{
+  "color": [
+    {
       "id": 1,
       "name": "Любые",
       "value": null
@@ -27,21 +31,25 @@
       "id": 2,
       "name": "Прозрачные",
       "value": "Прозрачные"
-    }, {
+    },
+    {
       "id": 3,
       "name": "Красные",
       "value": "Красные"
-    }, {
+    },
+    {
       "id": 4,
       "name": "Желтые",
       "value": "Желтые"
-    }, {
+    },
+    {
       "id": 5,
       "name": "Зеленые",
       "value": "Зеленые"
     }
   ],
-  "taste": [{
+  "taste": [
+    {
       "id": 1,
       "name": "Любые",
       "value": null
@@ -50,11 +58,13 @@
       "id": 2,
       "name": "Кислые",
       "value": "Кислые"
-    }, {
+    },
+    {
       "id": 3,
       "name": "Фруктовые",
       "value": "Фруктовые"
-    }, {
+    },
+    {
       "id": 4,
       "name": "Горькие",
       "value": "Горькие"
