+++ src/static/assets/json/dropdownItemSortBy.json
@@ -1,0 +1,24 @@
+{
+  "data": [
+    {
+      "id": "1",
+      "value": "price",
+      "name": "Цена"
+    },
+    {
+      "id": "2",
+      "value": "alco",
+      "name": "Крепость"
+    },
+    {
+      "id": "3",
+      "value": "name",
+      "name": "Название"
+    },
+    {
+      "id": "4",
+      "value": null,
+      "name": "По умолчанию"
+    }
+  ]
+}
