+++ src/store/modules/navbar/state.js
@@ -1,8 +1,9 @@
 
-import SortItem from './static/assets/json/dropdownItemSortBy.json';
-import FillterItem from './static/assets/json/dropdownItem.json';
+import SortItem from '../../../static/assets/json/dropdownItemSortBy.json';
+import FillterItem from '../../../static/assets/json/dropdownItem.json';
 
 export default {
+  serverUrl: 'http://localhost:3005',
   filterByAlco: '',
   filterByColor: '',
   filterByTaste: '',
